pyproject.toml
[project]
name = "quickstart"
version = "0.1.0"
description = "Soma scoring quickstart — run GPU scoring on Modal"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11",
    "modal>=1.3.4",
    "soma-models>=0.1.7",
    "soma-sdk>=0.1.7",
    "python-dotenv>=1.1",
    "boto3>=1.35",
]

[build-system]
requires = ["uv_build>=0.10.4,<0.11"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"

[dependency-groups]
dev = [
    "datasets>=4.6.1",
    "jax[cpu]>=0.9.1",
    "optax>=0.2.7",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "soma-models[all,flax]>=0.1.7",
]

[project.scripts]
settle-targets = "quickstart.settle_targets:main"
create-secrets = "quickstart.create_modal_secret:main"
stake-model = "quickstart.stake:main"
soma-status = "quickstart.status:main"
